AB#82798 Moved test util to separate folder CircleCI was thinking it was a test
